adapted to moved SimuComWorkflowConfiguration
